Add_a_gcp_sumologic_module_for_cloudlogging
This PR introduces a module for sending project wide logs from Google Cloud Platform to Sumologic.
I was not sure about where to put the examples. Normally they would
reside inside the module as a subfolder but I can move them to the main
"examples" folder.
The code has been tested and is actively being used in both GCP and
SumoLogic.